[Issue #179] [WIP] fixes tests
