+++ arability/app/views/search/search_with_filters.html.erb
@@ -266,7 +266,7 @@
         </span>
         <script type="text/javascript">
           <% if I18n.locale == :ar %>
-            $("#add_span").css("margin", "250px");
+            $("#add_span").css("margin", "230px");
           <% else %>
             $("#add_span").css("margin", "440px");
           <% end %>
