+++ arability/spec/controllers/projects_controller_khloud_spec.rb
@@ -1,0 +1,67 @@
+#encoding: UTF-8
+require "spec_helper"
+require "request_helpers"
+include Warden::Test::Helpers
+include RequestHelpers
+include Devise::TestHelpers
+
+describe ProjectsController, :type => :controller do
+
+  let(:word) {
+    word = Keyword.new
+    word.name = "testkeyword"
+    word.save
+    word
+  }
+
+  let(:syn) { syn = Synonym.new
+    syn.name = "كلمة"
+    syn.keyword_id = word.id
+    syn.save
+    syn
+  }
+
+  #khloud's tests
+
+  it "redirects to project path after calling export_to_csv if project empty", khloud: true do
+    p = create_project
+    get :export_to_csv, project_id: p.id
+    response.should redirect_to project_path(p.id)
+  end
+
+  it "redirects to project path after calling export_to_xml if project empty", khloud: true  do
+    p = create_project
+    get :export_to_xml, project_id: p.id
+    response.should redirect_to project_path(p.id)
+  end
+
+  it "redirectsto project path after calling export_to_json if project empty", khloud: true  do
+    p = create_project
+    get :export_to_json, project_id: p.id
+    response.should redirect_to project_path(p.id)
+  end
+
+  it "responds with ok code after calling export_to_csv with valid project", khloud: true  do
+    p = create_project
+    ps =
+    PreferedSynonym.add_keyword_and_synonym_to_project(syn.id, word.id, p.id)
+    get :export_to_csv, project_id: p.id
+    response.code.should eq("200")
+  end
+
+  it "responds with ok code after calling export_to_xml with valid project", khloud: true  do
+    p = create_project
+    ps =
+    PreferedSynonym.add_keyword_and_synonym_to_project(syn.id, word.id, p.id)
+    get :export_to_xml, project_id: p.id
+    response.code.should eq("200")
+  end
+
+  it "responds with ok code after calling export_to_json with valid project", khloud: true  do
+    p = create_project
+    ps =
+    PreferedSynonym.add_keyword_and_synonym_to_project(syn.id, word.id, p.id)
+    get :export_to_json, project_id: p.id
+    response.code.should eq("200")
+  end
+end
